Re-organize the site layout
- Re-organize App.tsx's layout
- Move all controls into a sidebar instead of mixed in with other
  components
