+++ src/App.module.scss
@@ -1,70 +1,38 @@
 .cubePlayground {
   margin: 0;
   display: grid;
-  grid-template-columns: repeat(4, auto);
-  grid-template-rows: repeat(7, auto);
+  grid-template-columns: 1fr repeat(2fr, 4);
+  grid-template-rows: 1fr repeat(2fr, 3) auto;
   grid-template-areas:
-    "size rset pick"
-    "cube cube cube"
-    "cube cube cube"
-    "cube cube cube"
-    "cube cube cube"
-    "cube cube cube"
-    "cube cube cube";
+    "head  head head head head"
+    "panel cube cube cube cube"
+    "panel cube cube cube cube"
+    "panel cube cube cube cube"
+    "foot  foot foot foot foot";
   row-gap: 5px;
   column-gap: 5px;
-  min-width: 800px;
+  min-width: 850px;
 
-  .resize,
-  .reset,
-  .styleSelect {
-    border: 1px solid black;
-    padding: 5px;
-    display: flex;
-    flex: row nowrap;
-    align-items: center;
-    justify-content: center;
-    gap: 5px;
-    max-width: 100%;
-
-    button {
-      padding: 5px;
-      width: 25px;
-      height: 25px;
-      background-color: #ffffff;
-      background-size: cover;
-      border-radius: 15%;
-      border: 1px solid transparent;
-      cursor: pointer;
-      transition: border-color 0.25s;
-
-      &:disabled {
-        background-color: #808080;
-      }
-    }
+  .head,
+  .controlPanel,
+  .cube,
+  .foot {
+    padding: 10px;
   }
 
-  .resize {
-    grid-area: size;
-
-    button {
-      background-image: url("/src/assets/progression.svg");
-    }
+  .head {
+    grid-area: head;
   }
 
-  .reset {
-    grid-area: rset;
-    button {
-      background-image: url("/src/assets/cycle.svg");
-    }
-  }
-
-  .styleSelect {
-    grid-area: pick;
+  .controlPanel {
+    grid-area: panel;
   }
 
   .cube {
     grid-area: cube;
-    padding: 10px;
+  }
+
+  .foot {
+    grid-area: foot;
   }
 }
