+++ src/components/cubes/rendered-cube.tsx
@@ -14,7 +14,6 @@ import React, { useMemo } from "react";
 import {
   FaceRotationButton,
   FocusFaceButton,
-  RotateCubeButtonContainer,
   SliceRotationButton,
 } from "../workflow/";
 import { CubeRenderStyle, IReactCubeProps } from "./generic-cube";
@@ -23,8 +22,6 @@ import styles, {
   back,
   bottom,
   cell,
-  control,
-  controlButton,
   cube,
   cubeContainer,
   cubePerspective,
@@ -80,12 +77,11 @@ export const RenderedCube: React.FC<IReactCubeProps> = props => {
 
   return (
     <>
-      <div className={`${cubeContainer} ${renderStyleClass}`} style={cssVars}>
-        <div className={control}>
-          <div className={controlButton}>
-            <RotateCubeButtonContainer dispatch={cubeDispatch} />
-          </div>
-        </div>
+      <div
+        className={`${cubeContainer} ${renderStyleClass}`}
+        style={cssVars}
+        tabIndex={0}
+      >
         <div className={perspectiveWrapper}>
           <div className={cubePerspective}>
             <div className={cube}>{...sides}</div>
