+++ src/components/workflow/index.ts
@@ -1,2 +1,3 @@
+export * from "./control-panel";
 export * from "./credits";
 export * from "./geometry-buttons";
